services/media/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docfliq-media"
version = "0.1.0"
description = "Docfliq Media Processing microservice — upload, transcode, image processing, secure URLs"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic[email]>=2.5",
    "pydantic-settings>=2.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "arq>=0.26",
    "python-jose[cryptography]>=3.3",
    "slowapi>=0.1.9",
    "aioboto3>=13.0",
    "Pillow>=10.0",
    "docfliq-shared",
]

[tool.uv.sources]
docfliq-shared = { path = "../../shared", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
